backend/requirements-win.txt
django==3.2.9
Pillow==9.1.0
djangorestframework==3.13.1
django-rest-knox==4.1.0
django-cors-headers==3.11.0
cmake==3.22.4
face-recognition==1.3.0
numpy==1.22.3